requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.36.0
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==11.1.0
